# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FoodHelper API
  version: v1
  description: API for managing products, recipes, and nutritional calculations.
servers:
  - url: /
    description: Current server
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/register:
    post:
      security: []
      tags:
        - Authentication
      summary: Register user
      description: Creates a user when the registration code is valid and returns a signed Bearer JWT.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              default:
                value:
                  username: elias
                  password: correct-horse-battery-staple
                  registrationCode: foodhelper-invite
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request or registration code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/auth/login:
    post:
      security: []
      tags:
        - Authentication
      summary: Log in
      description: Validates credentials and returns a signed Bearer JWT for Authorization headers.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              default:
                value:
                  username: elias
                  password: correct-horse-battery-staple
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/menus/{id}:
    get:
      tags:
        - Menus
      summary: Get menu
      description: Returns one menu snapshot with ordered days, nutritional totals, rule evaluation, stock summary, consumed stock, and missing products.
      operationId: getMenu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Menu returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
        '404':
          description: Menu not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Menus
      summary: Undo menu creation
      description: Returns every consumed quantity to its original stock entry, removes the linked money movement, and deletes the open menu so its planning can be established again. Closed menus cannot be undone.
      operationId: undoMenuCreation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Menu creation undone
        '400':
          description: Closed menu cannot be undone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Menu not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/menus/{id}/used-stock:
    get:
      tags:
        - Menus
      summary: List used stock for menu
      description: Returns the stock entries consumed when the menu was created, ordered by the original stock allocation.
      operationId: listMenuUsedStock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Consumed stock entries returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuUsedStockResponse'
        '404':
          description: Menu not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/menus/{id}/close:
    post:
      tags:
        - Menus
      summary: Close menu
      description: Closes a menu after its end date, saves statistics, and records an immutable snapshot for every selected person.
      operationId: closeMenu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseMenuRequest'
            example:
              personIds: [1, 2]
      responses:
        '200':
          description: Menu closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuStatsResponse'
        '400':
          description: Menu cannot be closed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Menu or selected person not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/menus/{id}/stats:
    get:
      tags:
        - Menus
      summary: Get menu stats
      description: Returns the saved period and month statistics captured when the menu was closed.
      operationId: getMenuStats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Menu stats returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuStatsResponse'
        '404':
          description: Menu stats not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid request body
        path:
          type: string
          example: /api/v1/products
        timestamp:
          type: string
          format: date-time
...
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - registrationCode
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 80
          description: Unique username used to log in
          example: elias
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: Plain password. It is hashed before storage.
          example: correct-horse-battery-staple
        registrationCode:
          type: string
          maxLength: 128
          description: Shared registration code required to create a user
          example: foodhelper-invite
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Registered username
          example: elias
        password:
          type: string
          description: User password
          example: correct-horse-battery-staple
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: Created or authenticated user identifier
          example: 1
        username:
          type: string
          description: Authenticated username
          example: elias
        accessToken:
          type: string
          description: JWT access token for Authorization headers
        tokenType:
          type: string
          description: Authorization scheme
          example: Bearer
        expiresAt:
          type: string
          format: date-time
          description: Instant when the access token expires
...
    NutrientRuleEvaluationResponse:
      type: object
      properties:
        value:
          type: number
          description: Observed average per planned day
        minimum:
          type: number
          nullable: true
        maximum:
          type: number
          nullable: true
        status:
          type: string
          enum: [BELOW_MINIMUM, WITHIN_RANGE, ABOVE_MAXIMUM, NOT_CONFIGURED]
    NutritionalRulesEvaluationResponse:
      type: object
      properties:
        plannedDays:
          type: integer
          format: int32
        calories:
          $ref: '#/components/schemas/NutrientRuleEvaluationResponse'
        carbohydrates:
          $ref: '#/components/schemas/NutrientRuleEvaluationResponse'
        proteins:
          $ref: '#/components/schemas/NutrientRuleEvaluationResponse'
        fats:
          $ref: '#/components/schemas/NutrientRuleEvaluationResponse'
...
    MenuResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Current menu identifier
          example: 1
        planningId:
          type: integer
          format: int64
          description: Planning identifier used to create the menu
          example: 5
        payerUserId:
          type: integer
          format: int64
          description: User identifier that assumes the menu period cost
          example: 1
        payerUsername:
          type: string
          description: Username that assumes the menu period cost
          example: elias
        startDate:
          type: string
          format: date
          description: First date covered by the menu period
          example: 2026-06-15
        endDate:
          type: string
          format: date
          description: Last date covered by the menu period
          example: 2026-06-22
        days:
          type: array
          description: Saved day menus in the menu period
          items:
            $ref: '#/components/schemas/PlanningDayResponse'
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
        stockSummary:
          $ref: '#/components/schemas/PlanningStockSummaryResponse'
        usedStock:
          type: array
          description: Stock entries consumed to satisfy the menu period
          items:
            $ref: '#/components/schemas/MenuUsedStockResponse'
        shoppingList:
          type: array
          description: Products still missing after consuming stock
          items:
            $ref: '#/components/schemas/MenuShoppingListItemResponse'
        nutritionalRules:
          $ref: '#/components/schemas/NutritionalRulesEvaluationResponse'
    CloseMenuRequest:
      type: object
      required: [personIds]
      properties:
        personIds:
          type: array
          minItems: 1
          uniqueItems: true
          description: People who followed the menu and receive an immutable history entry
          items:
            type: integer
            format: int64
          example: [1, 2]
...
    MenuStatsResponse:
      type: object
      properties:
        menuId:
          type: integer
          format: int64
          description: Menu identifier
          example: 1
        period:
          $ref: '#/components/schemas/MenuPeriodStatsResponse'
        month:
          $ref: '#/components/schemas/MenuPeriodStatsResponse'
    MenuPeriodStatsResponse:
      type: object
      properties:
        maxDay:
          $ref: '#/components/schemas/MenuPeriodStatsDayResponse'
        minDay:
          $ref: '#/components/schemas/MenuPeriodStatsDayResponse'
        averageCalories:
          type: number
          description: Average calories per day in the period
          example: 1450.25
        averageCarbohydrates:
          type: number
          description: Average carbohydrates per day in the period
          example: 120.5
        averageProteins:
          type: number
          description: Average proteins per day in the period
          example: 80.1
        averageFats:
          type: number
          description: Average fats per day in the period
          example: 45.75
        moneySpent:
          type: number
          description: Money spent in the period
          example: 32.4
    MenuPeriodStatsDayResponse:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Day date
          example: 2026-06-21
        calories:
          type: number
          description: Calories for that day
          example: 1450.25
    PlanningStockSummaryResponse:
      type: object
      properties:
        plannedDays:
          type: integer
          format: int32
          description: Days in the menu that already have at least one saved day
          example: 3
        distinctProducts:
          type: integer
          format: int32
          description: Distinct products used across all planned days
          example: 8
        calories:
          $ref: '#/components/schemas/PlanningStockSummaryCaloriesResponse'
        estimatedCost:
          type: number
          description: Total estimated cost covered by available stock
          example: 18.5
        requirements:
          type: array
          description: Per-product stock coverage requirements
          items:
            $ref: '#/components/schemas/PlanningStockRequirementResponse'
    PlanningStockSummaryCaloriesResponse:
      type: object
      properties:
        averagePerPlannedDay:
          type: number
          description: Average calories per planned day
          example: 1234.56
        maxDay:
          $ref: '#/components/schemas/PlanningStockSummaryDayCaloriesResponse'
        minDay:
          $ref: '#/components/schemas/PlanningStockSummaryDayCaloriesResponse'
    PlanningStockSummaryDayCaloriesResponse:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Day date
          example: 2026-06-15
        calories:
          type: number
          description: Calories for the day
          example: 1450.25
    PlanningStockRequirementResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        productName:
          type: string
          description: Product name
          example: Apple
        requiredUnits:
          type: number
          description: Total units required by the planning
          example: 4.5
        availableUnits:
          type: number
          description: Total units available in stock
          example: 6
        coveredUnits:
          type: number
          description: Units that can be covered by available stock
          example: 4.5
        missingUnits:
          type: number
          description: Units still missing after consuming all available stock
          example: 0
        estimatedCost:
          type: number
          description: Estimated cost for the covered units
          example: 12.75
    PlanningDayResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Planned day identifier
          example: 1
        date:
          type: string
          format: date
          description: Date of the planned day
          example: 2026-06-15
        sections:
          type: array
          description: Configured day parts selected for the day
          items:
            $ref: '#/components/schemas/PlanningSectionResponse'
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
    PlanningSectionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Section identifier
          example: 1
        dayPartId:
          type: integer
          format: int64
          description: Reusable day part identifier
          example: 1
        name:
          type: string
          description: Section name
          example: Lunch
        description:
          type: string
          description: Section description copied from the reusable day part
          example: Main meal of the day
        sortOrder:
          type: integer
          format: int32
          description: Explicit display order within the day
          example: 10
        products:
          type: array
          description: Ordered products consumed in this section
          items:
            $ref: '#/components/schemas/PlanningProductResponse'
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
...
    PlanningProductResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        productName:
          type: string
          description: Product name
          example: Apple
        units:
          type: number
          description: Consumed units
          example: 2
        grams:
          type: number
          description: Consumed grams
          example: 300
        sortOrder:
          type: integer
          format: int32
          description: Explicit display order within the section. Values are unique within the section.
          example: 10
        nutritionalValues:
          $ref: '#/components/schemas/NutritionalValuesResponse'
    MenuUsedStockResponse:
      type: object
      properties:
        stockEntryId:
          type: integer
          format: int64
          description: Stock entry identifier
          example: 7
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        productName:
          type: string
          description: Product name
          example: Apple
        usedUnits:
          type: number
          description: Quantity consumed from the stock entry
          example: 1.5
        price:
          type: number
          description: Price paid for one unit in this stock entry
          example: 2.49
        totalCost:
          type: number
          description: Total cost of the consumed quantity
          example: 3.74
        expirationDate:
          type: string
          format: date
          nullable: true
          description: Stock expiration date when available
          example: 2026-06-20
        entryDate:
          type: string
          format: date
          description: Date when the stock entry was received
          example: 2026-06-10
    MenuShoppingListItemResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Product identifier
          example: 1
        productName:
          type: string
          description: Product name
          example: Apple
        missingUnits:
          type: number
          description: Units still missing after consuming available stock
          example: 2
    NutritionalValuesResponse:
      type: object
      properties:
        calories:
          type: number
          description: Calories
          example: 52
        carbohydrates:
          type: number
          description: Carbohydrates
          example: 14
        proteins:
          type: number
          description: Proteins
          example: 0.3
        fats:
          type: number
          description: Fats
          example: 0.2